draw rectangle in CallFragment added getCanvas method in CallFragment draw message is received in MainActivity
TODO: //chat messages must be received in MainActivity

+++ app/src/main/java/com/example/prefixa_01/odgtest1/CanvasView.java
@@ -48,7 +48,7 @@ public class CanvasView extends View {
     protected void onDraw(Canvas canvas) {
         super.onDraw(canvas);
         Rectangle rectangle = Rectangle.get(getContext());
-        canvas.drawRect((float) rectangle.getXini(), (float) rectangle.getYini(), (float) rectangle.getXend(), (float) rectangle.getYend(), mPaint );
+        canvas.drawRect(rectangle.getXini(), rectangle.getYini(), rectangle.getXend(), rectangle.getYend(), mPaint );
     }
 
     public void clearCanvas() {
